# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Evron e-commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: search
    description: Search operations
  - name: editor
    description: Page editor operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: newArrival
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /search:
    get:
      operationId: searchProducts
      tags: [search]
      summary: Deep search products
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /stats:
    get:
      operationId: getStats
      tags: [products]
      summary: Get store stats summary
      responses:
        "200":
          description: Store stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /editor/config:
    get:
      operationId: getEditorConfig
      tags: [editor]
      summary: Get page editor configuration
      responses:
        "200":
          description: Editor config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorConfig"

    put:
      operationId: saveEditorConfig
      tags: [editor]
      summary: Save page editor configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorConfigInput"
      responses:
        "200":
          description: Saved config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorConfig"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Product:
      type: object
      required: [id, name, description, price, categoryId, colors, images, rating, stock, isFeatured, isNewArrival, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, description, price, categoryId, colors, images, stock]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean

    Category:
      type: object
      required: [id, name, description, image, productCount]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        image:
          type: string
        productCount:
          type: integer

    CategoryInput:
      type: object
      required: [name, description, image]
      properties:
        name:
          type: string
        description:
          type: string
        image:
          type: string

    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        image:
          type: string

    SearchResult:
      type: object
      required: [query, products, total]
      properties:
        query:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        suggestions:
          type: array
          items:
            type: string

    CartItem:
      type: object
      required: [productId, productName, price, quantity, selectedColor, image]
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        selectedColor:
          type: string
        image:
          type: string

    Cart:
      type: object
      required: [items, total, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer

    CartItemInput:
      type: object
      required: [productId, quantity, selectedColor]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        selectedColor:
          type: string

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    StoreStats:
      type: object
      required: [totalProducts, totalCategories, featuredCount, newArrivalsCount]
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        featuredCount:
          type: integer
        newArrivalsCount:
          type: integer

    EditorConfig:
      type: object
      required: [id, sections, updatedAt]
      properties:
        id:
          type: integer
        sections:
          type: array
          items:
            $ref: "#/components/schemas/EditorSection"
        updatedAt:
          type: string

    EditorSection:
      type: object
      required: [id, type, order, visible, data]
      properties:
        id:
          type: string
        type:
          type: string
        order:
          type: integer
        visible:
          type: boolean
        data:
          type: object
          additionalProperties: true

    EditorConfigInput:
      type: object
      required: [sections]
      properties:
        sections:
          type: array
          items:
            $ref: "#/components/schemas/EditorSection"