# lib/api-zod/src/generated/types/cart.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Evron e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}