# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Evron e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  productId: number;
  quantity: number;
  selectedColor: string;
}