# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Evron e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  description: string;
  image: string;
}