# lib/api-zod/src/generated/types/editorConfigInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Evron e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { EditorSection } from "./editorSection";

export interface EditorConfigInput {
  sections: EditorSection[];
}