# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Evron e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  colors: string[];
  images: string[];
  rating: number;
  reviewCount?: number;
  stock: number;
  isFeatured: boolean;
  isNewArrival: boolean;
  createdAt: string;
}