# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Evron e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface SearchResult {
  query: string;
  products: Product[];
  total: number;
  suggestions?: string[];
}